server/requirements.txt
mysql-connector-python
requests
flask
flask-cors
lxml
pywin32
python-pptx
exif